voice_note_summarizer/llm_version/requirements.txt
agent_framework
requests
python-dotenv
speech_recognition
pyaudio